requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pdfplumber==0.10.3
PyPDF2==3.0.1
pdfminer.six==20221105
plotly==5.18.0
reportlab==4.4.4
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
